requirements.txt
mcp
httpx
python-dotenv
upstox-python-sdk
requests
alpha-vantage